+++ CHANGELOG.md
@@ -37,3 +37,4 @@
 - Claws inflicts Bleeding for 5 seconds rather than 300
 - Fixed the duplicate message send when you get a Gene that you're missing the required Genes for
 - Players with Meaty can now shear themselves while sneaking
+- Wall Climbing is no longer an Attribute (why was it? What was I thinking?)